Search-3: UI-поиск в AccountWindowScene + SearchSmoke
AccountWindowScene: header списка писем обзавёлся TextField «Поиск»
(появляется, только если session.searchService != nil). Пока запрос пуст —
показываем messages из папки; как только непустой — displayedMessages
переключается на searchResults. moveSelection (↑/↓) работает с обоими
списками. Counter и «ищем…» индикатор — из session.isSearching.
SearchSmoke: 10 проверок QueryParser + LocalSearcher (invoice / from: /
is:unread / has:attachment / комбо / before: / is:flagged / пустой
запрос). Подключено к Scripts/smoke.sh.

+++ Packages/Storage/Sources/Storage/LocalSearcher.swift
@@ -51,7 +51,7 @@ public actor LocalSearcher {
             }
             if query.hasAttachment == true {
                 clauses.append("(m.flags & ?) != 0")
-                args.append(MessageFlags.hasAttachment.rawValue)
+                args.append(Int64(MessageFlags.hasAttachment.rawValue))
             }
             if let isUnread = query.isUnread {
                 if isUnread {
@@ -59,11 +59,11 @@ public actor LocalSearcher {
                 } else {
                     clauses.append("(m.flags & ?) != 0")
                 }
-                args.append(MessageFlags.seen.rawValue)
+                args.append(Int64(MessageFlags.seen.rawValue))
             }
             if query.isFlagged == true {
                 clauses.append("(m.flags & ?) != 0")
-                args.append(MessageFlags.flagged.rawValue)
+                args.append(Int64(MessageFlags.flagged.rawValue))
             }
             if let before = query.before {
                 clauses.append("m.date < ?")
